requirements.txt
fastapi
Pillow>=9.2.0
uvicorn[standard]